Move uncommitted files watcher into $lib/files
