+++ apps/desktop/src/lib/files/watcher.ts
